+++ packages/svelte/src/internal/client/dom/task.js
@@ -1,79 +1,85 @@
 import { run_all } from '../../shared/utils.js';
 
 // Fallback for when requestIdleCallback is not available
-export const request_idle_callback =
+const request_idle_callback =
 	typeof requestIdleCallback === 'undefined'
 		? (/** @type {() => void} */ cb) => setTimeout(cb, 1)
 		: requestIdleCallback;
 
-let is_micro_task_queued = false;
-let is_idle_task_queued = false;
+/** @type {Array<() => void>} */
+let boundary_micro_tasks = [];
 
 /** @type {Array<() => void>} */
-let queued_boundary_microtasks = [];
-/** @type {Array<() => void>} */
-let queued_post_microtasks = [];
-/** @type {Array<() => void>} */
-let queued_idle_tasks = [];
+let micro_tasks = [];
 
-export function flush_boundary_micro_tasks() {
-	const tasks = queued_boundary_microtasks.slice();
-	queued_boundary_microtasks = [];
+/** @type {Array<() => void>} */
+let idle_tasks = [];
+
+function run_boundary_micro_tasks() {
+	var tasks = boundary_micro_tasks;
+	boundary_micro_tasks = [];
 	run_all(tasks);
 }
 
-export function flush_post_micro_tasks() {
-	const tasks = queued_post_microtasks.slice();
-	queued_post_microtasks = [];
+function run_post_micro_tasks() {
+	var tasks = micro_tasks;
+	micro_tasks = [];
 	run_all(tasks);
 }
 
-export function flush_idle_tasks() {
-	if (is_idle_task_queued) {
-		is_idle_task_queued = false;
-		const tasks = queued_idle_tasks.slice();
-		queued_idle_tasks = [];
-		run_all(tasks);
-	}
+function run_idle_tasks() {
+	var tasks = idle_tasks;
+	idle_tasks = [];
+	run_all(tasks);
 }
 
-function flush_all_micro_tasks() {
-	if (is_micro_task_queued) {
-		is_micro_task_queued = false;
-		flush_boundary_micro_tasks();
-		flush_post_micro_tasks();
-	}
+function run_micro_tasks() {
+	run_boundary_micro_tasks();
+	run_post_micro_tasks();
 }
 
 /**
  * @param {() => void} fn
  */
 export function queue_boundary_micro_task(fn) {
-	if (!is_micro_task_queued) {
-		is_micro_task_queued = true;
-		queueMicrotask(flush_all_micro_tasks);
+	if (boundary_micro_tasks.length === 0 && micro_tasks.length === 0) {
+		queueMicrotask(run_micro_tasks);
 	}
-	queued_boundary_microtasks.push(fn);
+
+	boundary_micro_tasks.push(fn);
 }
 
 /**
  * @param {() => void} fn
  */
 export function queue_micro_task(fn) {
-	if (!is_micro_task_queued) {
-		is_micro_task_queued = true;
-		queueMicrotask(flush_all_micro_tasks);
+	if (boundary_micro_tasks.length === 0 && micro_tasks.length === 0) {
+		queueMicrotask(run_micro_tasks);
 	}
-	queued_post_microtasks.push(fn);
+
+	micro_tasks.push(fn);
 }
 
 /**
  * @param {() => void} fn
  */
 export function queue_idle_task(fn) {
-	if (!is_idle_task_queued) {
-		is_idle_task_queued = true;
-		request_idle_callback(flush_idle_tasks);
+	if (idle_tasks.length === 0) {
+		request_idle_callback(run_idle_tasks);
 	}
-	queued_idle_tasks.push(fn);
+
+	idle_tasks.push(fn);
+}
+
+/**
+ * Synchronously run any queued tasks.
+ */
+export function flush_tasks() {
+	if (boundary_micro_tasks.length > 0 || micro_tasks.length > 0) {
+		run_micro_tasks();
+	}
+
+	if (idle_tasks.length > 0) {
+		run_idle_tasks();
+	}
 }